+++ ChangeLog.md
@@ -1,3 +1,11 @@
+## 0.6.1
+
+* Switch to V2 upload by default
+
+## 0.6.0
+
+* Upload bundle V2 stuff
+
 ## 0.5.2
 
 * Upload LTS to Hackage with the name LTSHaskell